🚑️ Implement Comment model: add Comment type, sanitize function, and update related files. Remove Note model.
